requirements.txt
torch>=1.9.0
torchaudio>=0.9.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
PyYAML>=6.0
tqdm>=4.62.0
pathlib2>=2.3.6
librosa>=0.8.0
# AST specific dependencies
timm==0.4.5
wget>=3.2
